# openapi/cms-api.yaml
openapi: 3.1.0
info:
  title: CMS Collection Management System API
  version: 1.0.0
  description: |
    Multi-tenant Collection Management System API.
    
    ## Authentication
    All endpoints (except /health and /ready) require authentication via Bearer token.
    
    ## Multi-Tenancy
    All requests must include tenant context. In development, use the `x-tenant-id` header.
    In production, tenant is derived from the JWT token.
    
    ## Request Tracing
    Include `x-request-id` and `x-correlation-id` headers for distributed tracing.
  contact:
    name: CMS Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://api.cms.example.com
    description: Production server
...
paths:
...
  /api/v1/integration/runs:
    get:
      operationId: listIngestionRuns
      summary: List ingestion runs with filters
      tags:
        - Integration - Runs
      parameters:
        - name: dataset
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [running, success, partial, failed]
        - name: source_system
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of ingestion runs
  /api/v1/integration/runs/{run_id}:
    get:
      operationId: getIngestionRun
      summary: Get run with reconciliation summary
      tags:
        - Integration - Runs
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Run details with reconciliation
  /api/v1/integration/runs/{run_id}/items:
    get:
      operationId: getRunItems
      summary: Get individual items for a run
      tags:
        - Integration - Runs
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: outcome
          in: query
          schema:
            type: string
            enum: [INSERTED, UPDATED, SKIPPED, FAILED]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of ingestion items
  /api/v1/integration/runs/{run_id}/errors/csv:
    get:
      operationId: downloadRunErrorsCsv
      summary: Download errors as CSV file
      tags:
        - Integration - Runs
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: CSV file with error details
          content:
            text/csv:
              schema:
                type: string

  # ========== EPIC 5: Integration - Freshness ==========
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token with tenant and user claims
...
security:
  - bearerAuth: []